Add mesh event / command payloads.
This refactors the mesh-heartbeat into a generic event/mesh message,
which can be used for 'known' event and proprietary events. This also
adds a command/mesh message, which can be used for sending commands
to the Relay Gateways within the Gateway Mesh (this depends on using
the ChirpStack Gateway Mesh).